Encapsulated EDS readings logic
